Reclaim the phrase during "add" manipulation
Close #75
The phrase can be preserved when the "add" step in "modify" is failed.
After a new phrase is added, the selection of the original phrase in
chewing-editor seems to be reset, then the remove() in UserphaseView
will lost the remove target.
So I check the modified phrase add to the database successfully or not,
If the adding is fail, then re-add the original phrase.
See also: #109