Undo functionality for phrase deletion
The undo functionality can support multiple phrase deletion,
but users have to revert them one by one. When undo completed,
there will be a notification showing the phrase which users reverted.

I add a vector in UserphraseModel.cpp to record the removed phrase.
In this way can easily access the removed phrase vector in the remove
and undo method.

Resolves: #133