Refactor the parameter type of add function
Remove shared_ptr type for the QString buffer phrase and bopomofo,
because QString can implicitly share data which is same as shared_ptr.
Implicit sharing makes ownership of an object can be shared through
pointers and the object is copied only if and when the content is
going to be changed, i.e., copy-on-write.
Moreover, in UserphraseModel, I remove the public add() which only
calls the private add() and the private add() is the function really
add phrase to dictionary. Then modify the access specifier of the
add() from private to public.